move tests module into separate file
